requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
datasets>=2.12.0
opencv-python
pandas
scikit-learn
tqdm
matplotlib
pillow
